requirements.txt
streamlit
PyMuPDF
pytesseract
Pillow